Adjust options modal so that option name is not cut off
